third_party/WebKit/LayoutTests/inspector/sources/debugger/properties-special-expected.txt

CONSOLE MESSAGE: line 10: true
CONSOLE MESSAGE: line 11: function (a,b) { return a + b; }
CONSOLE MESSAGE: line 12: function () { [native code] }
Tests how debugger presents special properties of closures, bound functions and object wrappers.

properties-special.html:10 Boolean
    __proto__: Boolean
    [[PrimitiveValue]]: true
properties-special.html:11 function (a,b) { return a + b; }
    arguments: null
    caller: null
    length: 2
    name: ""
    prototype: Object
    __proto__: Empty() {}
    <function scope>
properties-special.html:12 function () { [native code] }
    arguments: (...)
    get arguments: ThrowTypeError() { [native code] }
    set arguments: ThrowTypeError() { [native code] }
    caller: (...)
    get caller: ThrowTypeError() { [native code] }
    set caller: ThrowTypeError() { [native code] }
    length: 1
    name: ""
    __proto__: Empty() {}
    [[TargetFunction]]: (a,b) { return a + b; }
    [[BoundThis]]: Object
    [[BoundArgs]]: Array[1]
